Move io to platform.
