+++ packages/yew/src/virtual_dom/vnode.rs
@@ -151,7 +151,7 @@ mod feat_ssr {
 
     use super::*;
     use crate::html::AnyScope;
-    use crate::io::BufWriter;
+    use crate::platform::io::BufWriter;
 
     impl VNode {
         pub(crate) fn render_into_stream<'a>(
