+++ packages/yew/src/virtual_dom/vsuspense.rs
@@ -28,7 +28,7 @@ impl VSuspense {
 mod feat_ssr {
     use super::*;
     use crate::html::AnyScope;
-    use crate::io::BufWriter;
+    use crate::platform::io::BufWriter;
     use crate::virtual_dom::Collectable;
 
     impl VSuspense {
